// posthog-core/src/posthog-surveys-types.ts
export interface SurveyAppearance {
  // keep in sync with frontend/src/types.ts -> SurveyAppearance
  backgroundColor?: string
  submitButtonColor?: string
  // deprecate submit button text eventually
  submitButtonText?: string
  submitButtonTextColor?: string
  ratingButtonColor?: string
  ratingButtonActiveColor?: string
  autoDisappear?: boolean
  displayThankYouMessage?: boolean
  thankYouMessageHeader?: string
  thankYouMessageDescription?: string
  thankYouMessageDescriptionContentType?: SurveyQuestionDescriptionContentType
  thankYouMessageCloseButtonText?: string
  borderColor?: string
  position?: 'left' | 'right' | 'center'
  placeholder?: string
  shuffleQuestions?: boolean
  surveyPopupDelaySeconds?: number
  // widget options
  widgetType?: 'button' | 'tab' | 'selector'
  widgetSelector?: string
  widgetLabel?: string
  widgetColor?: string
}

export enum SurveyType {
  Popover = 'popover',
  API = 'api',
  Widget = 'widget',
}

export type SurveyQuestion = BasicSurveyQuestion | LinkSurveyQuestion | RatingSurveyQuestion | MultipleSurveyQuestion

export type SurveyQuestionDescriptionContentType = 'html' | 'text'

interface SurveyQuestionBase {
  question: string
  description?: string | null
  descriptionContentType?: SurveyQuestionDescriptionContentType
  optional?: boolean
  buttonText?: string
  originalQuestionIndex: number
  branching?: NextQuestionBranching | EndBranching | ResponseBasedBranching | SpecificQuestionBranching
}

export interface BasicSurveyQuestion extends SurveyQuestionBase {
  type: SurveyQuestionType.Open
}

export interface LinkSurveyQuestion extends SurveyQuestionBase {
  type: SurveyQuestionType.Link
  link?: string | null
}

export interface RatingSurveyQuestion extends SurveyQuestionBase {
  type: SurveyQuestionType.Rating
  display: 'number' | 'emoji'
  scale: 3 | 5 | 7 | 10
  lowerBoundLabel: string
  upperBoundLabel: string
}

export interface MultipleSurveyQuestion extends SurveyQuestionBase {
  type: SurveyQuestionType.SingleChoice | SurveyQuestionType.MultipleChoice
  choices: string[]
  hasOpenChoice?: boolean
  shuffleOptions?: boolean
}
// ...
export type SurveyUrlMatchType = 'regex' | 'not_regex' | 'exact' | 'is_not' | 'icontains' | 'not_icontains'

export interface SurveyElement {
  text?: string
  $el_text?: string
  tag_name?: string
  href?: string
  attr_id?: string
  attr_class?: string[]
  nth_child?: number
  nth_of_type?: number
  attributes?: Record<string, any>
  event_id?: number
  order?: number
  group_id?: number
}
export interface SurveyRenderReason {
  visible: boolean
  disabledReason?: string
}

export interface Survey {
  // Sync this with the backend's SurveyAPISerializer!
  id: string
  name: string
  description: string
  type: SurveyType
  feature_flag_keys:
    | {
        key: string
        value?: string
      }[]
    | null
  linked_flag_key: string | null
  targeting_flag_key: string | null
  internal_targeting_flag_key: string | null
  questions: SurveyQuestion[]
  appearance: SurveyAppearance | null
  conditions: {
    url?: string
    selector?: string
    seenSurveyWaitPeriodInDays?: number
    urlMatchType?: SurveyUrlMatchType
    events: {
      repeatedActivation?: boolean
      values: {
        name: string
      }[]
    } | null
    actions: {
      values: SurveyActionType[]
    } | null
  } | null
  start_date: string | null
  end_date: string | null
  current_iteration: number | null
  current_iteration_start_date: string | null
}

export interface SurveyActionType {
  id: number
  name: string | null
  steps?: ActionStepType[]
}

/** Sync with plugin-server/src/types.ts */
export type ActionStepStringMatching = 'contains' | 'exact' | 'regex'

export interface ActionStepType {
  event?: string | null
  selector?: string | null
  /** @deprecated Only `selector` should be used now. */
  tag_name?: string
  text?: string | null
  /** @default StringMatching.Exact */
  text_matching?: ActionStepStringMatching | null
  href?: string | null
  /** @default ActionStepStringMatching.Exact */
  href_matching?: ActionStepStringMatching | null
  url?: string | null
  /** @default StringMatching.Contains */
  url_matching?: ActionStepStringMatching | null
}